- Added validation error for FE and BE

+++ frontend_react/src/page/UserPage.tsx
@@ -8,8 +8,10 @@ import {
 } from '../service/user/userApi';
 import EditUserDialog from '../component/user/edit/EditUserModal';
 import AddUserDialog from '../component/user/add/AddUserModal';
+import DeleteModal from '../component/user/delete/DeleteModal';
 import UserList from '../component/user/list/UserList';
 import { Button, Container, Stack, Typography } from '@mui/material';
+import Toast from '../component/template/Toast';
 
 interface User {
   id: number;
@@ -18,6 +20,12 @@ interface User {
 }
 
 const UserPage: React.FC = () => {
+  const [toastOpen, setToastOpen] = useState(false);
+  const [toastMessage, setToastMessage] = useState('');
+  const [toastSeverity, setToastSeverity] = useState<'success' | 'error'>(
+    'success'
+  );
+
   const { token } = useAuth()!;
   const [users, setUsers] = useState<User[]>([]);
   const [loading, setLoading] = useState<boolean>(true);
@@ -29,19 +37,42 @@ const UserPage: React.FC = () => {
   const [newUserEmail, setNewUserEmail] = useState<string>('');
   const [newUserPassword, setNewUserPassword] = useState<string>('');
   const [isAddModalOpen, setAddModalOpen] = useState(false);
-  const handleOpenAddModal = () => setAddModalOpen(true);
-  const handleCloseAddModal = () => setAddModalOpen(false);
+
   const [isEditModalOpen, setEditModalOpen] = useState(false);
+  const [userToDelete, setUserToDelete] = useState<User | null>(null);
+  const [isDeleteModalOpen, setDeleteModalOpen] = useState(false);
   const handleOpenEditModal = (user: User) => {
     setEditingUser(user);
     setUpdatedName(user.name);
     setUpdatedEmail(user.email);
     setEditModalOpen(true);
   };
+  const [formErrorSummary, setFormErrorSummary] = useState<string[]>([]);
+
+  const handleOpenAddModal = () => {
+    setAddModalOpen(true);
+    setFormErrorSummary([]);
+  };
+
+  const handleCloseAddModal = () => {
+    setAddModalOpen(false);
+    setFormErrorSummary([]);
+  };
 
   const handleCloseEditModal = () => {
+    setFormErrorSummary([]);
     setEditModalOpen(false);
     setEditingUser(null);
+  };
+
+  const handleOpenDeleteModal = (user: User) => {
+    setUserToDelete(user);
+    setDeleteModalOpen(true);
+  };
+
+  const handleCloseDeleteModal = () => {
+    setUserToDelete(null);
+    setDeleteModalOpen(false);
   };
 
   useEffect(() => {
@@ -63,11 +94,6 @@ const UserPage: React.FC = () => {
   }, [token]);
 
   const addUser = async () => {
-    if (!newUserName || !newUserEmail) {
-      setError('Name and Email are required');
-      return;
-    }
-
     try {
       const newUser = await addUserApi({
         name: newUserName,
@@ -80,9 +106,13 @@ const UserPage: React.FC = () => {
       setNewUserName('');
       setNewUserEmail('');
       setNewUserPassword('');
-    } catch (error) {
-      console.error('Error adding user:', error);
-      setError(error instanceof Error ? error.message : 'Something went wrong');
+
+      setToastOpen(true);
+      setToastMessage('Add new user success.');
+      setToastSeverity('success');
+      handleCloseAddModal();
+    } catch (error: any) {
+      setFormErrorSummary(Object.values(error.list));
     }
   };
 
@@ -103,26 +133,45 @@ const UserPage: React.FC = () => {
         )
       );
 
+      setFormErrorSummary([]);
       setEditingUser(null);
       setUpdatedName('');
       setUpdatedEmail('');
-    } catch (error) {
-      console.error('Error updating user:', error);
-      setError(error instanceof Error ? error.message : 'Something went wrong');
+
+      setToastOpen(true);
+      setToastMessage('Update success.');
+      setToastSeverity('success');
+      handleCloseEditModal();
+    } catch (error: any) {
+      console.log(error);
+      setFormErrorSummary(Object.values(error.list));
     }
   };
 
-  const handleDelete = async (userId: number) => {
+  const handleConfirmDelete = async () => {
+    if (!userToDelete) return;
+
     try {
       await deleteUserApi({
-        id: userId,
+        id: userToDelete.id,
         token,
       });
 
-      setUsers((prevUsers) => prevUsers.filter((user) => user.id !== userId));
+      setUsers((prevUsers) =>
+        prevUsers.filter((user) => user.id !== userToDelete.id)
+      );
+
+      setToastOpen(true);
+      setToastMessage('Delete success.');
+      setToastSeverity('success');
     } catch (error) {
+      setToastOpen(true);
+      setToastMessage('Delete error.');
+      setToastSeverity('error');
       console.error('Error deleting user:', error);
       setError(error instanceof Error ? error.message : 'Something went wrong');
+    } finally {
+      handleCloseDeleteModal();
     }
   };
 
@@ -136,6 +185,30 @@ const UserPage: React.FC = () => {
 
   return (
     <Container maxWidth="lg" sx={{ pt: 4, pb: 4 }}>
+      <Toast
+        open={toastOpen}
+        onClose={() => setToastOpen(false)}
+        message={toastMessage}
+        severity={toastSeverity}
+      />
+
+      <AddUserDialog
+        isAddModalOpen={isAddModalOpen}
+        newUserName={newUserName}
+        newUserEmail={newUserEmail}
+        newUserPassword={newUserPassword}
+        setNewUserName={setNewUserName}
+        setNewUserEmail={setNewUserEmail}
+        setNewUserPassword={setNewUserPassword}
+        handleCloseAddModal={handleCloseAddModal}
+        addUser={addUser}
+        formErrorSummary={formErrorSummary}
+      />
+
+      <Typography variant="h4" gutterBottom>
+        User List
+      </Typography>
+
       <Stack
         direction="row"
         spacing={2}
@@ -147,26 +220,18 @@ const UserPage: React.FC = () => {
           Add
         </Button>
       </Stack>
-      <AddUserDialog
-        isAddModalOpen={isAddModalOpen}
-        newUserName={newUserName}
-        newUserEmail={newUserEmail}
-        newUserPassword={newUserPassword}
-        setNewUserName={setNewUserName}
-        setNewUserEmail={setNewUserEmail}
-        setNewUserPassword={setNewUserPassword}
-        handleCloseAddModal={handleCloseAddModal}
-        addUser={addUser}
-      />
-
-      <Typography variant="h4" gutterBottom>
-        User List
-      </Typography>
 
       <UserList
         users={users}
         onEdit={handleOpenEditModal}
-        onDelete={handleDelete}
+        onDelete={handleOpenDeleteModal}
+      />
+
+      <DeleteModal
+        open={isDeleteModalOpen}
+        userName={userToDelete?.name || ''}
+        onClose={handleCloseDeleteModal}
+        onConfirm={handleConfirmDelete}
       />
 
       <EditUserDialog
@@ -178,8 +243,8 @@ const UserPage: React.FC = () => {
         onClose={handleCloseEditModal}
         onSave={() => {
           handleUpdate();
-          handleCloseEditModal();
         }}
+        formErrorSummary={formErrorSummary}
       />
     </Container>
   );
